Create LineIntersections.cpp
For this assignment, we were required to take in a list of input so that the first n instances represented nodes on a line, and the second list of instances represented nodes on a line just below that. Imaginatively, we could connect the 1st of the 1st n instances with the 1st of the 2nd n instances, the 2nd of the 1st n instances with the 2nd of the 2nd n instances, etc for all n instances. My job was to come up with an algorithm that counted the number of times these lines intersected. Incidentally, I noticed a pattern that was similar to last weeks assignment (inversion counting), where if you sort the first array (keeping the positions of the second array) and then sort the second array counting inversions, you will be left with the number of lines crossed. Thus, this is what I implemented for a runtime of ~2n log(n).