Dynamic Programming solution to interval scheduling problem
Code sorts an array and writes over it, incrementally storing max weight and the end length of that max at each position. Could have a better worst case (O(n log n) instead of O(n^2)) if when looking for k we would iterate through halves of 0 to j instead of just incrementing backwards, but this is a trivial step in my opinion.