Ford Fulkerson implementation of Network Flow
Algorithm finds max flow of a graph when given an edge list. It is a bit lengthy since one must repeatedly use BFS (or DFS) to find an augmenting path and then use the AP to update the original graph to a residual graph (see bottleneck function) until there are no more augmenting paths. As of right now there is some edge case in the final test that I am not passing, but this works for every input I've tried.